+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Switch migrateForm() param type to `EnvironmentTypes.FormMigrationData` and now return the migrated form
+
 ## [0.18.7] - 2022-01-19
 
 ### Added